==> 03_feature_engineering_and_encoding.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler
import time
import lightgbm as lgb
from lightgbm import early_stopping, log_evaluation

## <b>1.  Importing Data to Python</font></b>
---

feature_application_train = pd.read_csv('D:/Code Skripsi - RL and Ensemble/data/raw/outlier_application_train.csv')

feature_application_train.shape

feature_application_train.head()

## <b>2.  Financial Ratios</font></b>
---

# Financial Ratios

feature_application_train['CREDIT_INCOME_RATIO'] = feature_application_train['AMT_CREDIT'] / feature_application_train['AMT_INCOME_TOTAL']
feature_application_train['ANNUITY_INCOME_RATIO'] = feature_application_train['AMT_ANNUITY'] / feature_application_train['AMT_INCOME_TOTAL']
feature_application_train['CREDIT_ANNUITY_RATIO'] = feature_application_train['AMT_CREDIT'] / feature_application_train['AMT_ANNUITY']
feature_application_train['CREDIT_GOODS_RATIO'] = feature_application_train['AMT_CREDIT'] / feature_application_train['AMT_GOODS_PRICE']
feature_application_train['ANNUITY_GOODS_RATIO'] = feature_application_train['AMT_ANNUITY'] / feature_application_train['AMT_GOODS_PRICE']
feature_application_train['PAYMENT_CREDIT_RATIO'] = feature_application_train['AMT_PAYMENT'] / feature_application_train['AMT_CREDIT']

## <b>3.  Age & Employment Features</font></b>
---

# Age & Employment Features

feature_application_train['YEARS_EMPLOYED'] = feature_application_train['DAYS_EMPLOYED'] / -365
feature_application_train['YEARS_BIRTH'] = feature_application_train['DAYS_BIRTH'] / -365
feature_application_train['EMPLOYMENT_TO_AGE_RATIO'] = feature_application_train['YEARS_EMPLOYED'] / feature_application_train['YEARS_BIRTH']
feature_application_train['CREDIT_TO_EMPLOYMENT_RATIO'] = feature_application_train['AMT_CREDIT'] / (feature_application_train['YEARS_EMPLOYED'] + 1)
feature_application_train['CREDIT_TO_AGE_RATIO'] = feature_application_train['AMT_CREDIT'] / feature_application_train['YEARS_BIRTH']

## <b>4.  Family Responsibility Indicators</font></b>
---

# Family Responsibility Indicators

feature_application_train['INCOME_PER_FAMILY_MEMBER'] = feature_application_train['AMT_INCOME_TOTAL'] / (feature_application_train['CNT_FAM_MEMBERS'] + 1)
feature_application_train['CHILD_DEPENDENT_RATIO'] = feature_application_train['CNT_CHILDREN'] / (feature_application_train['CNT_FAM_MEMBERS'] + 1)
feature_application_train['HAS_CHILDREN'] = (feature_application_train['CNT_CHILDREN'] > 0).astype(int)

## <b>5.  Housing & Ownership Stability</font></b>
---


# # Housing & Ownership Stability
#
# feature_application_train['HAS_CAR_AND_REALTY'] = (feature_application_train['FLAG_OWN_CAR'] & feature_application_train['FLAG_OWN_REALTY']).astype(int)
# feature_application_train['NO_PROPERTY_OR_CAR'] = ((feature_application_train['FLAG_OWN_CAR'] == 0) & (feature_application_train['FLAG_OWN_REALTY'] == 0)).astype(int)
# feature_application_train['HOUSING_STABILITY'] = feature_application_train['YEARS_REGISTRATION'] / feature_application_train['YEARS_BIRTH']

feature_application_train.shape #NAMBAH LOAN TENURE (307511, 60) -> JADI (307511, 48) setelah seleksi kolom manual

# feature_application_train.to_csv("feature_application_train.csv", index=False)
